requirements.txt
# Core
langchain-core>=0.3.0
langgraph>=0.2.0
langchain-groq>=0.2.0
instructor>=1.5.0
pydantic>=2.0.0,<3.0.0
groq>=0.12.0

# Multi-provider LLM
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
openai>=1.0.0
anthropic>=0.40.0

# Data collection
httpx>=0.27.0
dnspython>=2.7.0
python-whois>=0.9.4

# Graph
networkx>=3.4
pyvis>=0.3.2

# Visualization / UI
streamlit>=1.40.0

# Export
fpdf2>=2.8.0
mistune>=3.0.0

# Utils
python-dotenv>=1.0.0

# Dev / Test
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
respx>=0.22.0
